Code improvements from review, added editor button
