--- Assets/Prefabs/Block.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: acb81bc73a0f7e74ba0c6e767e75feb9
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
