--- Assets/Prefabs/BloodsprayParticles.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a47e5bfb2d2cff94baf022374ed12a7c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
