--- Assets/Prefabs/Enemy.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a61d06283b9d62c459ae9b11b3dd1eec
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
