--- Assets/Scripts/HitScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 76d2b6da735cbfe43b7fbeabd72310bf
